initial fix to area issue, clearing when new boundary uploaded ...

+++ bfrs/api.py
@@ -184,17 +184,16 @@ class BushfireResource(APIResource):
     def hydrate_fire_boundary(self, bundle):
         if bundle.data.has_key('fire_boundary') and isinstance(bundle.data['fire_boundary'], list):
             bundle.data['fire_boundary'] = MultiPolygon([Polygon(*p) for p in bundle.data['fire_boundary']]).__str__()
-            bundle.obj.tenures_burnt.all().delete()
+
+            if bundle.data.has_key('area') and bundle.data['area'].has_key('tenure_area') and bundle.data['area']['tenure_area'].has_key('areas') and bundle.data['area']['tenure_area']['areas']:
+                bundle.obj.tenures_burnt.all().delete()
+
             if bundle.obj.report_status >= Bushfire.STATUS_INITIAL_AUTHORISED:
                 #import ipdb;ipdb.set_trace()
                 bundle.obj.final_fire_boundary = True
         else:
-#            if bundle.obj.report_status < Bushfire.STATUS_INITIAL_AUTHORISED:
-#                bundle.obj.initial_area = None
-#            else:
+#            if bundle.obj.report_status >= Bushfire.STATUS_INITIAL_AUTHORISED:
 #                bundle.obj.area = None
-            if bundle.obj.report_status >= Bushfire.STATUS_INITIAL_AUTHORISED:
-                bundle.obj.area = None
             bundle.obj.fire_boundary = None
             bundle.obj.final_fire_boundary = False
         return bundle
@@ -229,6 +228,7 @@ class BushfireResource(APIResource):
         elif bundle.data.has_key('tenure_ignition_point') and not bundle.data['tenure_ignition_point']:
             bundle.obj.tenure = Tenure.objects.get(name__iendswith='other')
 
+        #import ipdb;ipdb.set_trace()
         if bundle.data.has_key('area') and bundle.data['area'].has_key('tenure_area') and bundle.data['area']['tenure_area'].has_key('areas') and bundle.data['area']['tenure_area']['areas']:
             update_areas_burnt(bundle.obj, bundle.data['area']['tenure_area']['areas'])
